+++ jagobelanja-api/src/main/java/com/jagobelanja/service/UserService.java
@@ -1,7 +1,6 @@
 package com.jagobelanja.service;
 
 import com.jagobelanja.entity.User;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
@@ -10,7 +9,7 @@ public interface UserService {
     List<User> getUsers();
 
     //get user by id
-    User getUser(long id);
+    User getUser(int id);
 
     //create user
     User addUser(User user);
@@ -19,5 +18,5 @@ public interface UserService {
     User updateUser(User user);
 
     //delete user
-    void deleteUser(long id);
+    void deleteUser(int id);
 }
